Overall improvements

+++ AnimeAggregator.Models/AnimeUpdate.cs
@@ -8,7 +8,7 @@ namespace AnimeAggregator.Models
     {
         public Anime Anime { get; set; }
         public Publisher Publisher { get; set; }
-        public short EpisodeNum { get; set; }
+        public int EpisodeNum { get; set; }
         public string UpdateDate { get; set; }
         public DubType DubType { get;set;}
     }
